requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
passlib
bcrypt==3.2.0
python-jose[cryptography]